requirements.txt
django==2.1.2
whitenoise==4.1.2
gunicorn==19.9.0
Pillow==5.0.0